Improvements mostly concerning inventory.
o Added a collection of permissions for different actions related to inventory.
o Added ability to restrict inventory access to specified facilities and warehouses.
o Added the concept of "basic units" for a product template which is the number of inventory
  items in the ordered package. For example the facility might stock individual pills
  but sell them in bottles of 20, 50 or 100 with a separate template for each.
o Implemented "transaction types" for different kinds of actions taken on inventory.
  Improved related validation checks.
o Added validation and prompting for reason and notes when destroying lots.
o Added warehouse attribute to user facilities.
o Added patient ID as an index to the log table.
o Added check for missing encounter when dispensing inventory.
o Added some IPPF-specific logic related to contraception.
o Added ability to hide prices in the Fee Sheet.
o Support for negative units in the Fee Sheet, as when returning a previously sold item.
o Added an option for the default rendering provider.
o Added global settings specific to IPPF (will not appear for others).
o Assorted cleanup and bug fixes.