Initial FHIR Search implementation.
Built out the FHIR search algorithm and implemented it in
FhirServiceBase and BaseService.

I built the search upon a 3 layer approach.  The FHIR layer that is
aware of FHIR search fields types, FHIR concepts, the OpenEMR Service
layer, and then the database / SQL layer.  I believe I've kept those
layers separate from each other with the bottom layers not making any
calls or dependencies on the top layers.

The FHIR search uses the factory design pattern to build a set of
ISearchField objects that represent the various types of search objects
in both FHIR and now in the OpenEMR service layers.  Currently the type
of search fields that are supported are the String and Date fields which are
almost fully supported with Token kinda working and the others not
supported at this point. The only thing missing in the Date search
type is timezone support which I haven't got a good solution to just
yet.

I also implemented an IPatientCompartmentResourceService that any FHIR
service that touches patient data should implement.  It specifies the
patient field that will bind a FHIR search query into the patient
context.  This will make sure that only a single user's patient data
will be returned when operating within just that patient's context.

I've gone through and added a bunch of tests in the
FhirPatientServiceQueryTest object that deals with the different
modifiers, and comparators that FHIR supports.  I abstracted it away so
that the same kinds of modifiers and comparators (greater than, less
than, etc) can be used in the service classes without knowledge of FHIR.

I built it so we can have composite search fields with nested
heirarchies to support all kinds of search operations that can combine
both unions and intersections of search fields.

There is also a big bug fix as the sqlThrowException operation was only
working on insert/update operations that expected no data back.  I fixed
a bug that prevented the operation from returning data.