Code style fixes, unit test fixes.
Got all of the unit tests back up and running, had to fix a number of
problems with how the search parameters were defined and setup.  Also
fixed some broken tests with the Practitioner service classes.  Put more
logic into the BaseService to handle and be able to override table joins
and selection of data from table joins.

I've partially implemented a very primitive ORM, I keep wondering if itd
be better just to incorporate a pre-existing ORM but it seems like the
project has put those in and ripped them out multiple times so I'm
hedging away from that.

Had to temporarily hard code the Practitioner search values (which we
were doing originally anyways) until I can get the npi:missing stuff
working for a search modifier.  The challenge is that the :missing
identifier overrides the search value to be a boolean true/false which
is going to kill a number of the searchfield data validators.

I'm debating on whether I will under the hood change the type to be a
SearchMissingField class instead of relying on whatever the actual
search type of class should be.  I see pros and cons to that approach.
I'll just have to play around with it and see what I end up with.

Overall this should be a pretty good initial stab at a more robust FHIR
searching solution that will pass ONC requirements.