Patient Provenance and US-Core Patient requirements.
Got the Single Patient API ONC Us Core Inferno test suite to pass for
the US Core Patient Profile.  All of the core search criteria is working
as well as the gender, language, and race core extensions.  The
communication is throwing a warning but we use the same data values for
CCDA so I believe this is correct, but the test still passes despite the
warning.

Also implemented the Patient Provenance.  We'll see if the
implementation still works in the future as I'm not sure if our
Provenance record needs to contain EVERY single entity for that
Provenance target... we'll have to see as that could be pretty big at
the organization level.