Initial search work for FHIR Allergy Intolerance.
Created a query utils to put common query methods in that can be used by
the project that also throw exceptions and are easier to catch / handle
in the framework (IE returning an operation outcome if something goes
wrong rather than just dieing).

Extended FixtureManager to create fixtures for Allergy Intolerance.  Not
sure I like it as we want something more extensible.

Started the initial underpinnings to support the allergy intolerance
searching with FHIR.  We will be doing a reference search so we will
have that piece supported soon.