Fixed lookup codes if codeset is not installed
We had an error on the continuous integration service where the code
sets for allergy-intolerance are not installed.  Fixed it so that it
doesn't throw the error if the codeset is not found.  I'm not sure if
the FHIR unit tests just never checked against snomed codes before, but
I find it strange/bizarre that this never threw errors beforehand.