Force patient binding, fixed missing phone numbers.
Tracked why the patient fixtures would fail the test cases sometimes.
Since the patient is chosen at random in some of the patient fixtures it
was messing up on the phone records.  Found there was a patient that did
not have all of the phone numbers specified.  Fixing that resolves the
test cases that were failing.