Implement medical device, fix practitioner
Had some bugs in the practitioner resource.  Went ahead and fixed those
so they work properly with FHIR.  This ended up refactoring some of the
address and name elements into the FHIR utils class so we could remove
duplicate code in Patient, and Person.

Also implemented the DeviceService for searching on medical devices.