Implemented Goal&Careplan resources,Test Fixtures
Created a CarePlan service that handles the care plan form where goals and care plans are stored.  Since care plans do not have a unique id we had to implement a surrogate key using the encounter and the form_id of the care plan form.

Implemented the CodeTypesService that wraps around some of the global lookup code functions and provides additional helper method functions for grabbing the OID or system URI for a specific code.

Fixed some style issues in various classes.

Updated PatientService's getOne function to remove duplicate code.

Added some protection on npi number in the PractitionerService class to deal with the 'missing' modifier.

Heavily modified the test fixtures to be able to pull in referenced column data points for the fixture.  Wrote a small DSL to allow whitelisted function calls inside the JSON to do things like generate ids, and uuids.  Wrote a fixture manager for Encounters and care plan.