Implement faster UUID generation.
The current UUID implemention is very slow when large numbers of UUIDs
need to be populated.  This commit improves the performance for bulk
uuid generation dramatically.  Anything more than 1 record being
generated is faster with 5-10 records being 2.5X faster and anything
more than 100 records are seeing performance gains of 10X.

I needed this work for the vital forms generation as large clinics will
have a ton of vital form data that needs uuid generation.

There's room for a lot more optimization and tweaking but I'm putting
this out there for others to review / comment on and see if this new
approach will create problems.

In order to see the speed increases, we first precount the number of
uuids missing.  We then batch generate the ids and do a batch query to
make sure they are unique (generating new records if they are not).  We
then do a batch insert on the unique ids generated.  Everything is
rolled into a transaction so its idempotent for the generation.

On 0 records we do have the tradeoff that the operation now takes on
average 3ms instead of 0.96ms.  Single record performance is 20ms
compared to the previous 5.3ms.

The attached unit test shows the benchmarks up to 1000 records (which
the old system takes 5 seconds and the new one takes .5 seconds.