Fix user permissions and oauth2 fhirUser
The fhirUser for ONC certification needs to be a practitioner.  Made it
so that if a user with an NPI number logs in via the oauth2 interface
that their fhirUser returned in the JWT id_token is a Practitioner.  If
its a OpenEMR user that does not have an NPI number they are returned as
a Person.

Also added the missing scope permissions for the user/* context.