Refactored clinical notes, vitals to support FHIR
Added an auto incrementing primary key for form_clinical_notes.
Refactored the current id column to be its more accurately described form_id.  Changed the deletion
logic to actually set the activity status to be 0 like the regular forms table operates
(similar to procedures).  This fixes the ONC FHIR requirements that resources should not be deleted
and instead have their status property set to inactive or entered-in-error.

Refactored the form_clinical_notes form code to generate a uuid on save and creation. Moved a lot of
the sql logic into the ClinicalNotesService.  Also cleaned up some of the view code.

Made a minor fix in the vitals form table.sql file to have the table columns in.