Fix unique key on history_data
Now that we are using date and pid as the unique values that identify
the history_data we can add the unique key constraint.  Every record
should be unique unless a history_data is created at the exact same
timestamp for the same patient.  That shouldn't happen so we should be
good for this.