UuidRegistry make more DRY,Procedure UUIDs
Moved the table definitions for uuid population into a constant that can
now be used in all the other services.  This will simplify uuid
population / generation without duplicating the uuid table definition in
both the services class as well as the UuidRegistry. This makes the code
more DRY (Don't Repeat Yourself).

Also added uuids to the procedure_provider and procedure_report tables
in preparation for DiaganosticReport work.