Fix styles, primary organization reference.
Had to remove the FHIR id from the primary organization reference.
Having a reference element that is used in multiple places such as
DocumentReference where the organization is both the author and the
custodian makes the the FHIR validator fail as a reference entity's ID
must be unique (not the reference, but the entity id) which makes sense.
Removing the organization id fixes this problem.

Also fixed the code styling issues.