FHIRServiceBase cleanup, Organization, Encounter
Refactored some of the FHIRServiceBase so that we could remove a number of redundant getOne implementations in the various services.

Also implemented the Organization resource so we could get procedure providers, insurance companies, and facilities.  Made the searching work properly and changed up the Organization fhir service to be a facade similar to diagnostic report and document reference.

Fixed a number of problems in Encounter resource for FHIR