Openemr fhir diagnostic report (#4518)
* UuidRegistry make more DRY,Procedure UUIDs

Moved the table definitions for uuid population into a constant that can
now be used in all the other services.  This will simplify uuid
population / generation without duplicating the uuid table definition in
both the services class as well as the UuidRegistry. This makes the code
more DRY (Don't Repeat Yourself).

Also added uuids to the procedure_provider and procedure_report tables
in preparation for DiaganosticReport work.

* DiganosticReport for laboratory profile.

Implemented the diagnostic report for laboratory profile. Refactored the
queries in the procedure service and extended them so we can use them in
a couple of different contexts.

* Added procedure_provider to FHIR Organization

* Laboratory observation fixes.

* Fix value quantity for fhir resource

* fix rebase duplicate procedure type name

* Improve procedure service code w/ index

* FHIR DataAbsentReasons, Use standard code for labs

Fixed some of the code references on the dataAbsentReason which needs to
be a code type not a DataAbsentReason(DAR) extension.

Changed up the lab observations to use the LOINC standard code where
provided in order to make FHIR work.

* Renamed services for consistent naming

In order to follow consistent naming convention in other services (IE
DiagnosticReport) we rename observations to follow the same format.

* Fix Category, Custodian

* Fix styles, primary organization reference.

Had to remove the FHIR id from the primary organization reference.
Having a reference element that is used in multiple places such as
DocumentReference where the organization is both the author and the
custodian makes the the FHIR validator fail as a reference entity's ID
must be unique (not the reference, but the entity id) which makes sense.
Removing the organization id fixes this problem.

Also fixed the code styling issues.

* Partially fix organization unit tests

* FHIRServiceBase cleanup, Organization, Encounter

Refactored some of the FHIRServiceBase so that we could remove a number of redundant getOne implementations in the various services.

Also implemented the Organization resource so we could get procedure providers, insurance companies, and facilities.  Made the searching work properly and changed up the Organization fhir service to be a facade similar to diagnostic report and document reference.

Fixed a number of problems in Encounter resource for FHIR

* Fix encounter and provenance resources.

* Added typesafety to parseFhirResource.

Added types to the parseFhirResource in preparation to fix another bug.

* Typesafety for encounters,patients,practitioners.

Forced the resources to use their appropriate FHIR resource for insert
and post to provide more type safety and reduce some errors.  Fixed all
the unit tests to handle that correctly.

* Rename encounter search method to be consistent.

* Make exception message match if condition

* Remove duplicate comments.

* Rename FHIR Code System class

Since the code system class now has both URIs and OIDs for system's I'm renaming the class to be consistent with its purpose.

* Remove old todo

* Remove old TODO on social history

* Did the work and completing the todo

* Fix styles

* Add keys,indexes; phone #s to insurance.

Per code review suggestions, I added keys and indexes on the uuids into
the vanilla database script.  Also added phone numbers to the
organization FHIR endpoint for insurance companies.

Minor code cleanup as well.