Vitals Interpretation for FHIR #4527
This is a work in progress that will add clinician interpretations to the vital signs.  This current commit adds the interpretation value to the vital signs observations.  I'm currently mocking the values at the service level which provides a working example in FHIR for interpretation.  I'm still toying around with the database structure but the current implementation mocks things up nicely.  I add an interpretation_code and interpretation_text for every vital signs measurement.

I intended to overhaul the entire vitals_form but it is used in > 144 places with about half of those in the clinical decision rules.  So I'm taking this secondary step where we will have a vital_form_details table that will contain the interpretation and auxiliary data for each vitals_form.

I massively overhauled the vitals form UX in order to make the code less redundant and so I could easily add in the interpretation dropdown. I also reworked all of the metric / usa conversion codes on the vitals form to remove magic numbers.

The saving of the interpretation values is still not complete as well as the vital_form_details table implementation.  Overall though, I'm pretty pleased with how things turned out.