Form Vital Details, Refactored FormVitals,new list
Added a Form Vital Details concept to OpenEMR represented in the form_vital_details table and FormVitalDetails class.
This allows interpretations of the vitals to be stored for each vital and any additional information on a per column basis could be stored here.  While it'd be nice to breakup the form_vitals table itself and make it a vertical instead of horizontal representation, the form is used in many places that would be a pretty large overhaul at this point.  Future work of centeralizing all database interactions via the VitalsService would simplify this, but for now we will leverage the details representation.

Modified FormVitals to be populated via the vitals service as part of the initial work to centralize the DB interactions into the vitals service.

Refactored the FormVitals and C_FormVitals to be psr4 compliant.

Added a new list_options for the vitals-interpretations.

Did additional cleanup work on the form vitals.