FHIR Provenance Implementation.
I took a creative approach to implementing the provenance. Instead of
trying to figure out how to tie the provenance to each individual
resource I use the resource and the resource uuid to uniquely identify
the provenance resource.  Using that ID I can load the corresponding
FHIR resource service and re-run all of the queries and grab the correct
Provenance exactly as it was generated.

Since us-core doesn't require we list the provenance resources and we
just need to handle the read operation.

I had to modify the route parser so we could handle colon characters
inside the route.