Fixes #4560 Support multiple care team records.
Added a history table for the a patient's Care Team.  Any changes to the
care team providers or facilities that are part of the care team are
saved off to the patient_careteam_history table.  Note the status column
in patient_data is specifically for the current Care Team.  Past care
teams in the history table always have a status of 'inactive'.

I changed up the queries and report these changes back via FHIR.