Refactored ServiceLocator to prepare provenance.
In order to grab all of the provenance data we need to use our service
locator.  For that to happen we need to loop through all the services.
Since we have an export service locator that does this we refactored the
service locator and pulled it out into its own class that allows us to
find FHIR services that implement a particular class or sub-class.