Restrict PatientDocumentReference to patient docs
Had it so the patient doc reference was returning all documents instead
of just restricted to patient only.  This fixes a number of problems in
the bulk exporter.

However, system level documents (such as the exporter files themselves)
are not includes in the export now.  A future feature enhancement will
have to include these kinds of doc)