Client App access revoking
Implemented the ability to revoke a trusted user for a given registered
app.

Changed up the app interface to show more details about the registered
clients as well as all of the registered trusted users and any active
access tokens they may have at this point in time.

Put in the initial work to revoke an api authorization token.