Offline scope notice, Registration app type.
Added the ability to select what kind of app (confidential/public) to
register for.  I hide the system, user, and offline scopes if its a
public app.

I also add a notice per ONC requirements to users when the offline_scope
is requested to clearly show what is being granted and the
considerations there.