Fix unit tests for public refresh token.
Public clients do not get a refresh token and this was failing the unit
tests.  I added a test csae to the api test client tests to make sure we
ensure the refresh token is not sent back for public clients.

Removed the refresh token checks on introspection since a public client
can't have a refresh token.