Openemr fhir offline refresh scope (#4579)
* Offline scope notice, Registration app type.

Added the ability to select what kind of app (confidential/public) to
register for.  I hide the system, user, and offline scopes if its a
public app.

I also add a notice per ONC requirements to users when the offline_scope
is requested to clearly show what is being granted and the
considerations there.

* No offline_access disables refresh token

We weren't following the spec that prevents a refresh token from being
issued if the offline_access token is not provided.  I fixed the
authorization grants so that it handles the offline_access permission
set.

* Fix unit tests for public refresh token.

Public clients do not get a refresh token and this was failing the unit
tests.  I added a test csae to the api test client tests to make sure we
ensure the refresh token is not sent back for public clients.

Removed the refresh token checks on introspection since a public client
can't have a refresh token.

* Fix escaping and double space translations.