Remove forward slash escaping from JSON encoding
Our api endpoints were escaping the forward slashes in the JSON
responses.  This is optional in the JSON spec and is really only
necessary as far as I can tell when embedding JSON inside of an html
<script> tag.

However, some servers don't handle this forward slash escaping well.  In
the ONC Inferno test suite it does a strict string match for a URL.  The
match fails because of the escaped forward-slashes in the URL.  This is
why the data absent reason codes/extensions were failing.  It took many,
many hours to trace this down and fortunately its a simple fix.

This shouldn't effect any spec compliant clients who are consuming the
API as the forward slash escaping is an optional component and their
deserializers should still work.  Any non-spec compliant API client
could have issues.