Openemr single patient (#4599)
* MedicationRequest requestor if no practitioner

If the medication was created by someone who is not a practitioner we
have to put the reference as the organization... This usually only
happens if the administrator is not a clinician but adds a medication.

* Be able to report encounter on the prescription.

* Procedure codes, query fixes.

Had issues with the queries for procedures vs labs.  Had to add a
compound search clause to handle cases when results are empty which was
causing a binary string comparison to fail.

Also had to fix up the code formatting with the procedure codes.

* Fix styles

* Patient previous,suffix,careteam,ethnicity.

Fixed a number of patient issues.  I added support for the previous name
and suffix in FHIR.  Updated the database (should really only effect developers)
so that the history type key is set for careteams to leverage what Jerry
did for previous name.

Also fixed a problem with the ethnicity flags so that any unknown value
would exclude the FHIR ombCategory for ethnicity.

I had to change up the PatientService to override the search method.  I
did this to grab the previous name.  Standard api will now have a
previous name property.  Not sure I like giving the raw database values
back and those values are ignored on patient save right now.  I didn't
want to override what @sjpadgett did with patient name history so I left
that alone.

* Fix styles, fix uuid bug with patient service.

* Implement patient suffix parsing.