Group endpoints,optimize code lookups,Group export
Created grouping mechanism for patients to be grouped by their assigned provider.
Hitting the group endpoint lists each of the provider's and the patients assigned to those providers.  Only those with ACL admin.users permission can access

Added /Group and /Group/:groupId endpoints

Made it so the code lookups are fast for snomed codes.  This was needed for the Group export as just the condition resource alone was taking 5 seconds to return 5 resources.  Adding the lookups dropped this dramatically.

Made the group export use the group id to grab the patients that should be exported.  Now any FHIR resource that implements the IPatientCompartmentResourceService interface will extract ONLY the data that pertains to a specific group of patients.  Note we've tested this on a few patients but I have not seen what happens if the group is in the dozens to hundreds of members.