Set data absent if smoking status is missing.
Inferno validator was failing with a missing value concept.  Instead of
populating dataAbsent on the elemtn I instead put the data absent into
the valueCodeableConcept.  The spec 'prefers' the smoking codes from
VSAC but I couldn't look it up in the VSAC database and so I'm using HL7's null
value system for the missing data reason.