Reject invalid JWT aud and launch parameters.
SMART on FHIR spec requires that we inject an invalid 'aud' parameter
that does not point to the FHIR server.  I added this to the smart app
pieces.

SMART on FHIR spec requires that we inject an invalid 'launch' parameter
that does not originate from the server.  Rather than saving the launch
into the database I encrypt the encoded values.  If we can decrypt the
values properly we say the launch param comes from the server.  We could
add some HMAC signature here but this should be fine as everything must
be verified via the oauth2 login anyways.  If we implement the 'remember
me' functionality we will need to go through and add this HMAC piece in.