Bulk Export, Security Fixes, Group endpoint support (#4600)
* Client Credentials Security Fixes

Prevent client credentials grant JWT replay attacks.  Also found that we
weren't using the enabled flag for client credential grants.  Fixed that
so that only enabled system clients can actually make requests against
the api.

* Fix styles

* Group endpoints,optimize code lookups,Group export

Created grouping mechanism for patients to be grouped by their assigned provider.
Hitting the group endpoint lists each of the provider's and the patients assigned to those providers.  Only those with ACL admin.users permission can access

Added /Group and /Group/:groupId endpoints

Made it so the code lookups are fast for snomed codes.  This was needed for the Group export as just the condition resource alone was taking 5 seconds to return 5 resources.  Adding the lookups dropped this dramatically.

Made the group export use the group id to grab the patients that should be exported.  Now any FHIR resource that implements the IPatientCompartmentResourceService interface will extract ONLY the data that pertains to a specific group of patients.  Note we've tested this on a few patients but I have not seen what happens if the group is in the dozens to hundreds of members.

* Implemented group patient export for services.

Added the group patient export to the other US Core services that allow
filtering by the patient.

* Caught uncaught fatal error exception.

Invalid argument exceptions were thrown if a field's values were
missing.  Changed these to search field exceptions to make sure the
system catches them.

* Fix missing name columns for practitioner

I made the practitioner match what shows in the UX admin screen.  If
there is no username and npi number we don't consider them a user.

* Fix whitespace on care plan.

* Set data absent if smoking status is missing.

Inferno validator was failing with a missing value concept.  Instead of
populating dataAbsent on the elemtn I instead put the data absent into
the valueCodeableConcept.  The spec 'prefers' the smoking codes from
VSAC but I couldn't look it up in the VSAC database and so I'm using HL7's null
value system for the missing data reason.

* More text trimming to validate encounter resources.

* Fix patient author reference.

* Only return careteam if we have data.

* Fixed name fields for organization.

If we don't have an organization name we can't display anything which we
fixed in the procedure provider organizations.

We also make sure to trim the narrative for the name fields in all the
organizations.

* Add missing use statement.

Not sure why this use statement went missing but adding it back in.

* Reject invalid JWT aud and launch parameters.

SMART on FHIR spec requires that we inject an invalid 'aud' parameter
that does not point to the FHIR server.  I added this to the smart app
pieces.

SMART on FHIR spec requires that we inject an invalid 'launch' parameter
that does not originate from the server.  Rather than saving the launch
into the database I encrypt the encoded values.  If we can decrypt the
values properly we say the launch param comes from the server.  We could
add some HMAC signature here but this should be fine as everything must
be verified via the oauth2 login anyways.  If we implement the 'remember
me' functionality we will need to go through and add this HMAC piece in.

* Fix test trimming on goals

* Organization npi is must support not required

* Fix style fixes

* Fix practitioner,client credential tests.

Had a bug in the practitioner search for patching.  Also had a bug with
the unit tests for client credentials which I had to go and fix.

* More practitioner unit test fixes

* Fix styles.

* Add uuid mapping, fix launch encryption.

standardEncrypt/standardDecrypt already handles base64 encoding so I
remove the redundant step.  I also fix some other code review items
where we need to make sure we auto generate the uuid mapping for Groups
and change some datetime columns to be consistent with OpenEMR database.

I removed some old commented out code for encounters as well.