Scope Event, Api helper methods/logs
Added an event to handle adding additional scopes via event listeners.
If a module adds to the REST api the api is pretty useful if there is
not an accompanying scope added with the api.

Also added some helper methods to the http request object and some
more api logging.  I needed this for a module I am building and it
will be useful for the community to have it in core.