Fixes #4658 change uuid mapping to be on save.
Created a FHIR module that listens to the vitals, patient, and social
history save events and populates the uuid mapping tables whenever those
records are saved.  This removes the populate uuid mechanisms from being
triggered on every single api requests and instead the uuids are
populated at the time of save.

Added events to the social history service that triggers whenever a
history data record is saved.